remove database record of deleted files

+++ app/Http/Controllers/FileController.php
@@ -40,14 +40,16 @@ class FileController extends Controller
     public function delete(Request $request){
       $filename = $request->filename;
       $find_file = DB::table('files')->where('filename', '=', $filename)->first();
+      if (!$find_file){
+        return response()->json(['status' => 'file not found']);
+      }
       $hashed_filename = $find_file->hashed_filename;
       $exists = Storage::disk('local')->exists($hashed_filename);
       if ($exists){
         Storage::disk('local')->delete($hashed_filename);
+        DB::table('files')->where('filename', '=', $filename)->delete();
         return response()->json(['status' => 'successfully deleted file']);
-      } else {
-        return response()->json(['status' => 'file not found']);
-      }
+      } 
     }
 
     public function show(Request $request){
